Create README.md
This is a function template class. Each time this function is executed, it will first check whether the same parameter has been calculated. If it has been calculated, it will directly return the previously saved calculation result. Otherwise, it will be calculated first, and the parameters and results will be stored in the cache. This toy is based on template metaprogramming.